refactor(core): extract repo HTTP endpoints into feature router (#433)
Implements #425.2 / Group B of the ADR-015 core audit. Two
repo-feature-layer endpoints moved out of `server/hub-node-router.ts`
into a new `server/features/repo-router.ts`:

- GET  /hub/repo-inventory
- POST /hub/nodes/:nodeId/sessions/reopen  (cold reopen flow)

Pure routing (pairing, heartbeat envelope dispatch, node list,
node lifecycle, direct sessions create) stays in
`hub-node-router.ts`.

Changes:

- server/hub-node-router.ts: export shared utilities used by both
  routers (errorStatus, sendRelayError, sendRegistryError, relayError,
  isSessionSummary, scopedNodeSession, sessionFromPayload, bodyRecord,
  recordField, findColdReopenTarget, coldReopenSessionPayload,
  ColdReopenWarning, ColdReopenTarget). Delete the two repo endpoints.
- server/features/repo-router.ts (new): owns the two repo endpoints.
  Imports shared helpers from hub-node-router. RepoFeatureRouterOptions
  mirrors the relevant subset of HubNodeRouterOptions.
- server/index.ts: composition root now mounts BOTH routers, sharing
  one repoInventoryFeature + one collectLocalRepoInventory closure.
- test/hub-node-cold-transfer.test.ts: mount createRepoFeatureRouter
  alongside createHubNodeRouter (cold-reopen route lives there now).
- test/hub-node-routes.test.ts: mount createRepoFeatureRouter in the
  one test that hits /hub/repo-inventory.

37/37 tests across cold-transfer, routes, session-routing,
multi-node-smoke, repo-inventory-feature, production-wiring.

Closes #433.
